requirement.txt
audioread==3.0.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==2.1.1
cmake==3.25.0
contourpy==1.1.0
cycler==0.11.0
decorator==5.1.1
filelock==3.9.0
fonttools==4.42.1
fsspec==2023.9.0
gensim==4.3.2
huggingface-hub==0.16.4
idna==3.4
imageio==2.31.3
importlib-resources==6.0.1
Jinja2==3.1.2
joblib==1.3.2
kiwisolver==1.4.5
lazy_loader==0.3
librosa==0.10.1
lit==15.0.7
llvmlite==0.40.1
loguru==0.7.1
MarkupSafe==2.1.2
matplotlib==3.7.2
mpmath==1.2.1
msgpack==1.0.5
networkx==3.0
numba==0.57.1
numpy==1.24.1
packaging==23.1
pandas==2.1.0
Pillow==9.3.0
platformdirs==3.10.0
pooch==1.7.0
pycparser==2.21
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2023.3.post1
PyWavelets==1.4.1
PyYAML==6.0.1
regex==2023.8.8
requests==2.28.1
ruamel.yaml==0.17.32
ruamel.yaml.clib==0.2.7
safetensors==0.3.3
scikit-image==0.21.0
scikit-learn==1.3.0
scipy==1.11.2
seaborn==0.12.2
six==1.16.0
smart-open==6.3.0
soundfile==0.12.1
soxr==0.3.6
sympy==1.11.1
threadpoolctl==3.2.0
tifffile==2023.8.30
tokenizers==0.13.3
torch==2.0.1+cu118
torchaudio==2.0.2+cu118
torchlibrosa==0.1.0
torchvision==0.15.2+cu118
tqdm==4.66.1
transformers==4.33.0
triton==2.0.0
typing_extensions==4.4.0
tzdata==2023.3
urllib3==1.26.13
zipp==3.16.2
pycocoevalcap==1.2
sentence-transformers==2.2.2
